requirements.txt
selenium
beautifulsoup4
psycopg2-binary
psycopg2
logzero
python-dateutil

